refactor: relocate tools and add alert hook
