+++ README.md
@@ -1235,10 +1235,10 @@ ignored.
 ### 6.4 · Web Dashboard Quick-Start 📊
 Launch the local web interface:
 ```bash
-uvicorn src.interface.api_server:app --reload
-streamlit run src/interface/web_app.py
+uvicorn alpha_factory_v1.demos.alpha_agi_insight_v1.src.interface.api_server:app --reload
+streamlit run alpha_factory_v1/demos/alpha_agi_insight_v1/src/interface/web_app.py
 # React client
-cd src/interface/web_client
+cd alpha_factory_v1/demos/alpha_agi_insight_v1/src/interface/web_client
 npm ci          # use the lock file for reproducible installs
 npm run dev       # http://localhost:5173
 # build production assets
@@ -1250,13 +1250,13 @@ Alternatively run inside Docker:
 # build the web client first so `dist/` exists
 make build_web
 # regenerate protobuf modules and Go stubs
-./tools/gen_proto_stubs.sh  # updates src/utils/a2a_pb2.py and tools/go_a2a_client/a2a.pb.go
+./tools/gen_proto_stubs.sh  # updates alpha_factory_v1/core/utils/a2a_pb2.py and tools/go_a2a_client/a2a.pb.go
 make compose-up  # builds and waits for healthy services
 ```
 Run `./tools/gen_proto_stubs.sh` whenever `src/utils/a2a.proto` changes to keep the
 Python and Go stubs up to date.
 Open <http://localhost:8080> in your browser. When `RUN_MODE=web`, the container
-serves the static files from `src/interface/web_client/dist` using `python -m
+serves the static files from `alpha_factory_v1/demos/alpha_agi_insight_v1/src/interface/web_client/dist` using `python -m
 http.server`. The FastAPI demo also mounts this folder at `/` when present so the
 dashboard is reachable without additional tooling.
 
@@ -1272,7 +1272,7 @@ complexity from the final population.
 
 If Streamlit isn't installed or you're running on a headless server, use:
 ```bash
-python -m src.interface.minimal_ui --text
+python -m alpha_factory_v1.demos.alpha_agi_insight_v1.src.interface.minimal_ui --text
 ```
 to display the forecast results directly in the console.
